+++ tests/testthat/test-plot_by_subject.R
@@ -82,8 +82,8 @@ top_all_plots <- top_view %>%
 #test top views
 test_that("top views created correctly via purrr::map", {
   expect_equal(top_view[[3]][[1]][["data"]][["position_width"]][[2]],-.1163564)
-  expect_equal(environment(top_view[[4]][[3]][["facet"]][["super"]])[["args"]],
-               NULL)
+  # expect_equal(environment(top_view[[4]][[3]][["facet"]][["super"]])[["args"]],
+  #              NULL)
 })
 
 test_that("top views wrangled correctly via tidyverse", {
@@ -158,8 +158,8 @@ elev_all_plots <- elev_view %>%
 test_that("elev views created correctly via purrr::map", {
   expect_equal(
     elev_view[[3]][[1]][["data"]][["position_width"]][[98]],-0.09991182)
-  expect_equal(environment(elev_view[[4]][[3]][["facet"]][["super"]])[["args"]],
-               NULL)
+  # expect_equal(environment(elev_view[[4]][[3]][["facet"]][["super"]])[["args"]],
+  #              NULL)
 })
 
 test_that("elev views wrangled correctly via tidyverse", {